requirements.txt
asgiref==3.4.1
Django==3.2.8
djangorestframework==3.12.4
flake8==4.0.1
flake8-builtins==1.5.3
flake8-import-order==0.18.1
flake8-print==4.0.0
mccabe==0.6.1
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
